Add files via upload
Adding files used for exploratory analysis of speaker classification w/ Naive Bayes (bag of words) approach. This approach is not incredibly accurate at speaker classification when given 80% of the dialogue for training and tested on 20% of the dialogue (max of 60% in testing). I suspect this is because speakers naturally repeat words to each other in dialogue, so the bag of words assumption is insufficient. Later investigations might employ artificial words appended to each sentence to denote code-switching or will seek to classify speaker groups between dialogues rather than speakers within the dialogue. For example, if A and B are in dialogue 1, and C and D are in dialogue 2, train the model to differentiate between dialogue 1 and 2.